Fix windows compilation

+++ sql/field.h
@@ -1385,7 +1385,7 @@ public:
     }
     return update_fl;
   }
-  virtual void store_field_value(uchar *val, uint len)
+  virtual void store_field_value(uchar *val, size_t len)
   {
      memcpy(ptr, val, len);
   }
@@ -5088,7 +5088,7 @@ public:
     }
     return update_fl;
   }
-  void store_field_value(uchar *val, uint) override
+  void store_field_value(uchar *val, size_t) override
   {
     store(*((longlong *)val), TRUE);
   }
